[Front] Create Relationship component
